+++ src/BizHawk.Client.EmuHawk/tools/TAStudio/HeaderEditor.cs
@@ -7,10 +7,10 @@ namespace BizHawk.Client.EmuHawk
 {
 	public partial class MovieHeaderEditor : Form
 	{
-		private readonly IMovie _movie;
+		private readonly IBasicMovieInfo _movie;
 		private readonly Config _config;
 
-		public MovieHeaderEditor(IMovie movie, Config config)
+		public MovieHeaderEditor(IBasicMovieInfo movie, Config config)
 		{
 			_movie = movie;
 			_config = config;
